requirements.txt
djangorestframework
pymysql
chatbot-py